pyproject.toml
[build-system]
requires = ["setuptools >= 64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dqml_app"
version = "0.0.1"
authors = [
  { name="Rajakumaran Arivumani", email="author@example.com" },
]
description = "Data quality ML app package"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
#license = "MIT"
#license-files = ["LICEN[CS]E*"]
dependencies = [
  # app specific 
  "great-expectations == 1.3.0",
  "pandas == 2.1.4",
  "numpy", 
  "datetime",
  "xgboost == 2.1.4",
  "scikit-learn[model_selection]",
  "scikit-learn[pre_processing]",
  "shap == 0.46.0",
  "matplotlib == 3.10.0",
  "python-dotenv", 
  # my packages
  "utils",
  "metadata",
  "app_calendar",
  "config",
  # my packages
  #"utils @ git+https://github.com/dexplorer/utils",
  #"metadata @ git+https://github.com/dexplorer/df-metadata",
  #"app_calendar @ git+https://github.com/dexplorer/df-app-calendar",
  #"config @ git+https://github.com/dexplorer/df-config",
]
keywords = ["data", "quality", "etl", "elt", "framework", "governance"]

[project.optional-dependencies]
api = [
  "fastapi" , 
  "uvicorn[standard]", 
  ]
cli = [
  "click",
]
test = [
  # standard dev tools
  "pytest",
  "pytest-cov",
  "pylint",
  "black",
  "ipython",
]

[project.scripts]
dqml-app-cli = "dqml_app.dqml_app_cli:main"
dqml-app-api = "dqml_app.dqml_app_api:main"

[project.urls]
Homepage = "https://profile.rajakumaran.in"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/dexplorer/df-data-quality-ml.git"
Issues = "https://github.com/dexplorer/df-data-quality-ml/issues"
Changelog = "https://github.com/dexplorer/df-data-quality-ml/blob/master/CHANGELOG.md"
